requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.18
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cloudinary==1.36.0
fastapi-cors==0.0.6
firebase-admin==6.5.0
alembic==1.13.1
slowapi==0.1.9
